Assignment-2/Problem_Statement_1/Python/Method_2/pyproject.toml
[project]
name = "method-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "requests>=2.32.4",
    "python-dotenv>=1.0.0",
]
